refactor(reco-core): one rig basis - view_matrix folds onto rig_frame in the geometry leaf
Step 7 part 2: VirtualCamera and the rig correction move into
src/geometry, and view_matrix now composes its tilted+rolled reference
frame through the same rig_frame the pose inverse and the viewport-roll
margins use. The construction the comment used to say 'must mirror
view_matrix exactly' IS view_matrix's construction now - the three
sites can no longer drift. Guards: the world_to_render_pose round-trip
test, the view-matrix self-consistency test, and the CPU/GPU agreement
oracle (incl. the tilt+roll-at-yaw case) all pass unchanged.
